Move join_node preview testing to dedicated cpp
