+++ test/CMakeLists.txt
@@ -408,6 +408,7 @@ if (TARGET TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_join_node DEPENDENCIES TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_join_node_key_matching DEPENDENCIES TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_join_node_msg_key_matching DEPENDENCIES TBB::tbb)
+    tbb_add_test(SUBDIR tbb NAME test_join_node_preview DEPENDENCIES TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_limiter_node DEPENDENCIES TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_priority_queue_node DEPENDENCIES TBB::tbb)
     tbb_add_test(SUBDIR tbb NAME test_queue_node DEPENDENCIES TBB::tbb)
