+++ test/tbb/test_join_node.cpp
@@ -22,10 +22,10 @@
 #include "common/config.h"
 
 #include "test_join_node.h"
+#include "common/test_join_node_multiple_predecessors.h"
 
 //! \file test_join_node.cpp
 //! \brief Test for [flow_graph.join_node] specification
-
 
 static std::atomic<int> output_count;
 
@@ -122,147 +122,6 @@ public:
     }
 };
 
-#if __TBB_PREVIEW_FLOW_GRAPH_NODE_SET
-#include <array>
-#include <vector>
-void test_follows_and_precedes_api() {
-    using msg_t = tbb::flow::continue_msg;
-    using JoinOutputType = std::tuple<msg_t, msg_t, msg_t>;
-
-    std::array<msg_t, 3> messages_for_follows = { {msg_t(), msg_t(), msg_t()} };
-    std::vector<msg_t> messages_for_precedes = {msg_t(), msg_t(), msg_t()};
-
-    follows_and_precedes_testing::test_follows
-        <msg_t, tbb::flow::join_node<JoinOutputType>, tbb::flow::buffer_node<msg_t>>(messages_for_follows);
-    follows_and_precedes_testing::test_follows
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::queueing>>(messages_for_follows);
-    follows_and_precedes_testing::test_follows
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::reserving>, tbb::flow::buffer_node<msg_t>>(messages_for_follows);
-    auto b = [](msg_t) { return msg_t(); };
-    class hash_compare {
-    public:
-        std::size_t hash(msg_t) const { return 0; }
-        bool equal(msg_t, msg_t) const { return true; }
-    };
-    follows_and_precedes_testing::test_follows
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::key_matching<msg_t, hash_compare>>, tbb::flow::buffer_node<msg_t>>
-        (messages_for_follows, b, b, b);
-
-    follows_and_precedes_testing::test_precedes
-        <msg_t, tbb::flow::join_node<JoinOutputType>>(messages_for_precedes);
-    follows_and_precedes_testing::test_precedes
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::queueing>>(messages_for_precedes);
-    follows_and_precedes_testing::test_precedes
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::reserving>>(messages_for_precedes);
-    follows_and_precedes_testing::test_precedes
-        <msg_t, tbb::flow::join_node<JoinOutputType, tbb::flow::key_matching<msg_t, hash_compare>>>
-        (messages_for_precedes, b, b, b);
-}
-#endif
-
-namespace multiple_predecessors {
-
-using namespace tbb::flow;
-
-using join_node_t = join_node<std::tuple<continue_msg, continue_msg, continue_msg>, reserving>;
-using queue_node_t = queue_node<std::tuple<continue_msg, continue_msg, continue_msg>>;
-
-void twist_join_connections(
-    buffer_node<continue_msg>& bn1, buffer_node<continue_msg>& bn2, buffer_node<continue_msg>& bn3,
-    join_node_t& jn)
-{
-    // order, in which edges are created/destroyed, is important
-    make_edge(bn1, input_port<0>(jn));
-    make_edge(bn2, input_port<0>(jn));
-    make_edge(bn3, input_port<0>(jn));
-
-    remove_edge(bn3, input_port<0>(jn));
-    make_edge  (bn3, input_port<2>(jn));
-
-    remove_edge(bn2, input_port<0>(jn));
-    make_edge  (bn2, input_port<1>(jn));
-}
-
-std::unique_ptr<join_node_t> connect_join_via_make_edge(
-    graph& g, buffer_node<continue_msg>& bn1, buffer_node<continue_msg>& bn2,
-    buffer_node<continue_msg>& bn3, queue_node_t& qn)
-{
-    std::unique_ptr<join_node_t> jn( new join_node_t(g) );
-    twist_join_connections( bn1, bn2, bn3, *jn );
-    make_edge(*jn, qn);
-    return jn;
-}
-
-#if TBB_PREVIEW_FLOW_GRAPH_FEATURES
-std::unique_ptr<join_node_t> connect_join_via_follows(
-    graph&, buffer_node<continue_msg>& bn1, buffer_node<continue_msg>& bn2,
-    buffer_node<continue_msg>& bn3, queue_node_t& qn)
-{
-    auto bn_set = make_node_set(bn1, bn2, bn3);
-    std::unique_ptr<join_node_t> jn( new join_node_t(follows(bn_set)) );
-    make_edge(*jn, qn);
-    return jn;
-}
-
-std::unique_ptr<join_node_t> connect_join_via_precedes(
-    graph&, buffer_node<continue_msg>& bn1, buffer_node<continue_msg>& bn2,
-    buffer_node<continue_msg>& bn3, queue_node_t& qn)
-{
-    auto qn_set = make_node_set(qn);
-    auto qn_copy_set = qn_set;
-    std::unique_ptr<join_node_t> jn( new join_node_t(precedes(qn_copy_set)) );
-    twist_join_connections( bn1, bn2, bn3, *jn );
-    return jn;
-}
-#endif // TBB_PREVIEW_FLOW_GRAPH_FEATURES
-
-void run_and_check(
-    graph& g, buffer_node<continue_msg>& bn1, buffer_node<continue_msg>& bn2,
-    buffer_node<continue_msg>& bn3, queue_node_t& qn, bool expected)
-{
-    std::tuple<continue_msg, continue_msg, continue_msg> msg;
-
-    bn1.try_put(continue_msg());
-    bn2.try_put(continue_msg());
-    bn3.try_put(continue_msg());
-    g.wait_for_all();
-
-    CHECK_MESSAGE(
-        (qn.try_get(msg) == expected),
-        "Unexpected message absence/existence at the end of the graph."
-    );
-}
-
-template<typename ConnectJoinNodeFunc>
-void test(ConnectJoinNodeFunc&& connect_join_node) {
-    graph g;
-    buffer_node<continue_msg> bn1(g);
-    buffer_node<continue_msg> bn2(g);
-    buffer_node<continue_msg> bn3(g);
-    queue_node_t qn(g);
-
-    auto jn = connect_join_node(g, bn1, bn2, bn3, qn);
-
-    run_and_check(g, bn1, bn2, bn3, qn, /*expected=*/true);
-
-    remove_edge(bn3, input_port<2>(*jn));
-    remove_edge(bn2, input_port<1>(*jn));
-    remove_edge(bn1, *jn); //Removes an edge between a sender and port 0 of a multi-input successor.
-    remove_edge(*jn, qn);
-
-    run_and_check(g, bn1, bn2, bn3, qn, /*expected=*/false);
-}
-} // namespace multiple_predecessors
-
-
-#if __TBB_PREVIEW_FLOW_GRAPH_NODE_SET
-//! Test follows and precedes API
-//! \brief \ref error_guessing
-TEST_CASE("Test follows and preceedes API"){
-    test_follows_and_precedes_api();
-}
-#endif
-
 //! Test hash buffers behavior
 //! \brief \ref error_guessing
 TEST_CASE("Tagged buffers test"){
@@ -289,8 +148,4 @@ TEST_CASE("Test removal of the predecessor while having none") {
     using namespace multiple_predecessors;
 
     test(connect_join_via_make_edge);
-#if TBB_PREVIEW_FLOW_GRAPH_FEATURES
-    test(connect_join_via_follows);
-    test(connect_join_via_precedes);
-#endif
 }
